% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{s2_authors}
\alias{s2_authors}
\title{Semantic Scholar Author Lookup}
\usage{
s2_authors(s2_author_id = NULL, .progress = "none")
}
\arguments{
\item{s2_author_id}{character, Semantic Scholar Author ID.}

\item{.progress}{Shows the \code{plyr}-style progress bar.
Options are "none", "text", "tk", "win", and "time".
See \code{\link[plyr]{create_progress_bar}} for details
of each. By default, no progress bar is displayed.}
}
\description{
Obtain metadata relative to authors from Semantic Scholar including
links to papers, citation profile and name variants.
Please play nice with the API. If you need to access
more data, there is also a data dump available.
For more info see \url{https://api.semanticscholar.org/corpus/}.
}
\examples{
\dontrun{
# one author
s2_authors("2204561")

# multiple authors at once with progress bar
s2_authors(c("2204561", "144128278", "49930593"), .progress = "text")
}
}
