% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_papers.R
\name{s2_papers}
\alias{s2_papers}
\title{Lookup metadata about an article in Semantic Scholar}
\usage{
s2_papers(ids = NULL, .progress = "none")
}
\arguments{
\item{ids}{character vector, get metadata by a single identifier or many
identifiers. Besides its own id scheme, Semantic Scholar supports
DOIs and arXiv IDs.}

\item{.progress}{Shows the \code{plyr}-style progress bar.
Options are "none", "text", "tk", "win", and "time".
See \code{\link[plyr]{create_progress_bar}} for details
of each. By default, no progress bar is displayed.}
}
\description{
Obtain information about individual scholarly articles indexed in Semantic
Scholar. Please play nice with the API. If you need to access
more data, there is also a data dump available.
For more info see \url{https://api.semanticscholar.org/corpus/}.
}
\examples{
\dontrun{
get_papers("10.1186/1471-2164-11-245")
# call many papers with different id types at once
get_papers(ids = c("10.1093/nar/gkr1047",
"bbc25a6a340365832d4d27f683646c39f2661c88",
"10.7717/peerj.2323",
"arXiv:0711.0914"))
}

}
